Freelance/NLP_TASK/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
google-api-core==1.30.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.20.0
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.7.1
google-cloud-language==2.1.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==2.10
libcst==0.3.19
mypy-extensions==0.4.3
numpy==1.21.0
oauthlib==3.1.1
packaging==20.9
pandas==1.2.5
pandas-gbq==0.15.0
proto-plus==1.18.1
protobuf==3.17.3
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
